LayoutTests/http/tests/inspector/stacktraces/csp-injected-content-warning-contains-stacktrace-expected.txt

CONSOLE MESSAGE: Refused to load the image 'http://127.0.0.1:8000/resources/abe.png' because it violates the following Content Security Policy directive: "img-src 'none'".

This test injects a blocked image from JavaScript. The resulting console error should contain a stack trace.


Message[0]: csp-injected-content-warning-contains-stacktrace.html:12 Refused to load the image 'http://127.0.0.1:8000/resources/abe.png' because it violates the following Content Security Policy directive: "img-src 'none'".

Stack Trace:

    0) http://127.0.0.1:8000/inspector/stacktraces/csp-injected-content-warning-contains-stacktrace.html / thisTest / 12 / 0
    1) (internal script) / eval code / (line number) / (column number)
    2) [native code] / eval / 0 / 0
    3) [native code] / evaluate / 0 / 0
    4) (internal script) / _evaluateOn / (line number) / (column number)
    5) (internal script) / _evaluateAndWrap / (line number) / (column number)
    6) (internal script) / evaluate / (line number) / (column number)
